src/test/java/com/example/dailyplanner/service/UserServiceImplTest.java
package com.example.dailyplanner.service;

import com.example.dailyplanner.model.User;
import com.example.dailyplanner.repository.UserRepository;
import org.junit.jupiter.api.Test;
import org.mockito.Mockito;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.security.crypto.password.PasswordEncoder;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.mockito.Mockito.doReturn;

@SpringBootTest
public class UserServiceImplTest {
    @Autowired
    private UserServiceImpl userService;

    @MockBean
    private UserRepository userRepository;

    @MockBean
    private PasswordEncoder passwordEncoder;

    @Test
    public void testCreate() throws Exception {
        User user = new User();
        user.setPassword("Qwerty123!");
        String encodedPassword = "jf2w0jisdlfnweijdf0qjiedsolkfmwepofdjmsd";
        doReturn(encodedPassword).when(passwordEncoder)
                .encode(user.getPassword());
        userService.saveOrUpdate(user);
        Mockito.verify(userRepository).save(user);
        assertEquals(encodedPassword, user.getPassword());
    }
}
